core/message: introduce OpenProtocolMessage
The incoming message has to be parsed based on the
header and later on created specific MID class.

The OpenProtocolMessage is kind of raw message where
the header is parsed and payload kept as it is.
The payload will be parsed in the upper layer - the
concrete MID message.

Tests: add basic parsing test of defined/not defined
items.

Signed-off-by: Tomas Krcka <tomas.krcka@industware.cloud>