core: introduce a basic message and codec
The incoming message is first decoded to
OpenProtocolMessage. This message is then
validated whether the system knows the MID or not.

Introduce abstract class OpenProtocolMid, the
class will know how to process the payload and
how to encode to payload.

MidCodec uses these classes to recognize whether the
system knows it or not.

Signed-off-by: Tomas Krcka <tomas.krcka@industware.cloud>