application/client: extend client for loop
to be able to wait for a response, subscribed events.

The openprotocol has more communications flows.
It can be only sent and forget, send and wait a response,
send a subscription message and receiving event.

The _loop will be used for that reason, receive any
message must always be based on initial action
from an integrator. It means the message is always expected.

If not, log a warning for now.

The unittest were updated accordingly.

Signed-off-by: Tomas Krcka <tomas.krcka@industware.cloud>