core/messages: add the ability to set raw_str bytes
The RawMessage has already an ability to get bytes
 from the provided byte or bytes range.
When we'd like to test decoding messages received from
the controller, the message has to be created first.

Add an ability to set exact bytes in raw_str based on
a specification.
Note: setting raw_str doesn't set payload.
If the payload is necessary then decode() has to be
called after the modification.

Signed-off-by: Tomas Krcka <tomas.krcka@industware.cloud>