application: add Tightening application MIDs
The Tightening application is Event-based.
It contains a Subscribe, Event, Ack and Unsubscribe message.

The event data can have multiple revisions, revision 1
is different to rev2 and higher, only a few bytes are similar.

The parsing is focused on rev1 and rev2 data, anything in
higher revision is not parsed at the moment.

Adding unit-tests as well.

Signed-off-by: Tomas Krcka <tomas.krcka@industware.cloud>