core: not use revision for class registration
The MID can have multiple revisions and the definition
in the class covered only one.
To avoid creating multiple classes or have list of revisions,
don't use it for that, and the decoding will raise an
exception if the revision is not supported.

The revision will only hold the revision of the openmessage.
Add the default contructor to always set the REVISION information.

Signed-off-by: Tomas Krcka <tomas.krcka@industware.cloud>